Refactor: useEnterLoginFirst
